requirements.txt
# ==============================================================================
# Mistri.AI – Phase 2 RAG Pipeline Requirements
# ==============================================================================

# --- Web Framework ---
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.6.0

# --- AWS SDK (Bedrock: Titan Embeddings + Claude 3.5 Sonnet) ---
boto3==1.34.34
botocore==1.34.34

# --- Local Vector DB (FAISS – mocks Amazon Aurora locally) ---
faiss-cpu==1.7.4

# --- LangChain (text chunking utilities only) ---
langchain==0.1.6
langchain-community==0.0.19

# --- Image & PDF Processing ---
Pillow==10.2.0
PyMuPDF==1.23.26         # fitz – fast PDF-to-image extraction

# --- Data Science Stack ---
numpy==1.26.3
pandas==2.2.0

# --- Configuration & Utilities ---
python-dotenv==1.0.0
httpx==0.26.0            # async HTTP for health-check tests

# --- Testing ---
pytest==8.0.0
pytest-asyncio==0.23.5
